M1-Challenge-Rivera-Nelson/CustomerComposition/src/main/java/com/company/Customer.java
package com.company;

public class Customer {

    private String firstName;
    private String lastName;
    private boolean rewardsMember;


    // Getters
    public String getFirstName() {
        return firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public boolean isRewardsMember() {
        return rewardsMember;
    }


    // Setters
    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public void setRewardsMember(boolean rewardsMember) {
        this.rewardsMember = rewardsMember;
    }
}
